requirements.txt
# Core dependencies
loguru>=0.7.0
python-telegram-bot>=20.0
pyyaml>=6.0

# Async support
aiohttp>=3.9.0

# Data handling
python-decimal>=1.0

# Configuration
python-dotenv>=1.0.0

# Testing (optional)
pytest>=7.4.0
pytest-asyncio>=0.21.0
